Update main menu
