+++ Assets/Scripts/Main Menu/MainMenu.cs
@@ -53,16 +53,5 @@ public class MainMenu : MonoBehaviour
         _animator.SetTrigger("CharacterBack");
         ShowMainMenu();
     }
-    public void OpenSteamPage()
-    {
-        Application.OpenURL("https://store.steampowered.com/app/1949310/Eternals_Path/");
-    }
-    public void OpenDiscordPage()
-    {
-        Application.OpenURL("https://discord.gg/YvBYf74Paj");
-    }
-    public void OpenLinktreePage()
-    {
-        Application.OpenURL("https://linktr.ee/sleeplessparadise");
-    }
+  
 }
