+++ Assets/Scripts/Ui/OpenSocialLink.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1d00fd11a6acab446a6314f4cfdae0ce
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
